% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{add_emissions}
\alias{add_emissions}
\title{Add emissions settings to a model}
\usage{
add_emissions(
  model,
  rate = NULL,
  duration = NULL,
  start_day = NULL,
  start_hour = NULL,
  name = NULL
)
}
\arguments{
\item{model}{A model object created by `create_disp_model()`.}

\item{rate}{Emissions rate.}

\item{duration}{Emissions duration.}

\item{start_day}{Start day for emissions.}

\item{start_hour}{Start hour for emissions.}

\item{name}{Emissions name (defaults to "emissions_1", "emissions_2", ...).}
}
\value{
Updated model object.
}
\description{
Add emissions settings to a model
}
