% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{add_species}
\alias{add_species}
\title{Add a species definition to a model}
\usage{
add_species(
  model,
  name = NULL,
  pdiam = NULL,
  density = NULL,
  shape_factor = NULL,
  ddep_vel = NULL,
  ddep_mw = NULL,
  ddep_a_ratio = NULL,
  ddep_d_ratio = NULL,
  ddep_hl_coeff = NULL,
  wdep_hl_coeff = NULL,
  wdep_in_cloud = NULL,
  wdep_below_cloud = NULL,
  rad_decay = NULL,
  resuspension = NULL
)
}
\arguments{
\item{model}{A model object created by `create_disp_model()`.}

\item{name}{Species name (defaults to "species_1", "species_2", ...).}

\item{pdiam}{Particle diameter.}

\item{density}{Particle density.}

\item{shape_factor}{Shape factor.}

\item{ddep_vel}{Dry deposition velocity.}

\item{ddep_mw}{Dry deposition molecular weight.}

\item{ddep_a_ratio}{Dry deposition A ratio.}

\item{ddep_d_ratio}{Dry deposition D ratio.}

\item{ddep_hl_coeff}{Dry deposition Henry's law coefficient.}

\item{wdep_hl_coeff}{Wet deposition Henry's law coefficient.}

\item{wdep_in_cloud}{Wet deposition in-cloud coefficient.}

\item{wdep_below_cloud}{Wet deposition below-cloud coefficient.}

\item{rad_decay}{Radioactive decay.}

\item{resuspension}{Resuspension factor.}
}
\value{
Updated model object.
}
\description{
Add a species definition to a model
}
