% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{create a set of directories to run disperseR}
\usage{
create_dirs(location = file.path("~", "Desktop"))
}
\arguments{
\item{location}{Path where project directories are created. Defaults to the
Desktop when available, otherwise the home directory. Creates the path
if it does not exist.}
}
\value{
Creates directories (does not overwrite if existing). Outputs string variables with paths to the environment.
}
\description{
`create_dirs()` takes 0 or the following argument `location`
}
\details{
\code{create_dirs}
}
