% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{create_disp_model}
\alias{create_disp_model}
\title{Create a dispersion model object}
\usage{
create_disp_model(name = NULL)
}
\arguments{
\item{name}{Optional model name.}
}
\value{
A list with class `disp_model`.
}
\description{
Create a dispersion model object
}
