% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid of lat/lon points}
\usage{
create_grid(lat, lon, range, division)
}
\arguments{
\item{lat}{Center latitude}

\item{lon}{Center longitude}

\item{range}{Range around center (lat, lon)}

\item{division}{Grid spacing (lat, lon)}
}
\value{
List with lat and lon vectors
}
\description{
Creates a regular grid of latitude/longitude points.
This is a local implementation based on SplitR::create_grid.
}
\keyword{internal}
