% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_impact_table_single.R
\name{create_impact_table_single}
\alias{create_impact_table_single}
\title{Create impact table for single unit/month}
\usage{
create_impact_table_single(
  data.linked,
  data.units,
  link.to = "zips",
  zcta.dataset = NULL,
  counties. = NULL,
  map.month,
  map.unitID,
  metric = "N"
)
}
\arguments{
\item{data.linked}{Data from disperseR::link_all_units()}

\item{data.units}{Unit location information from disperseR::units()}

\item{link.to}{Spatial scale: 'zips', 'counties', or 'grids'}

\item{zcta.dataset}{ZCTA spatial dataset from disperseR::get_data()}

\item{counties.}{US counties sf object}

\item{map.month}{Month in YYYYMM format}

\item{map.unitID}{Unit ID string matching 'ID' column in data.units}

\item{metric}{Metric column name. Defaults to 'N'}
}
\value{
sf data.table of impacts for plotting
}
\description{
Creates a table of spatial impacts from linked data for 
a single unit and month, suitable for plotting.
}
