% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_env.R
\name{disperseR_cache_clear}
\alias{disperseR_cache_clear}
\title{Clear the disperseR session cache}
\usage{
disperseR_cache_clear()
}
\value{
Invisibly returns NULL.
}
\description{
Removes all cached directory paths and datasets. You will need
  to call `create_dirs()` again after clearing.
}
