% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{get_output_df}
\alias{get_output_df}
\title{Extract output from a model object}
\usage{
get_output_df(model)
}
\arguments{
\item{model}{A `traj_model` or `disp_model` list.}
}
\value{
A data.frame containing model output, or an empty data.table with
  columns (particle_no, lon, lat, height, hour) if absent.
}
\description{
Extract output from a model object
}
\keyword{internal}
