% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearmon.R
\name{get_yearmon}
\alias{get_yearmon}
\title{Generate a year-month sequence}
\usage{
get_yearmon(
  start.year = NULL,
  start.month = NULL,
  end.year = NULL,
  end.month = NULL
)
}
\arguments{
\item{start.year}{what year do you want your vector to start with? eg. start.year = "2005"}

\item{start.month}{what month of `start.year` eg. start.month = "01"}

\item{end.year}{what year do you want your vector to end with? This has to be a character eg. end.year = "2005"}

\item{end.month}{what month of that year? eg. end.month = "03"}
}
\value{
A character vector of year-month strings in \code{YYYYMM} format.
}
\description{
`get_yearmon()` outputs a character vector of year-month
  strings (format \code{YYYYMM}) that you can loop over. For January
  through March 2005, use `start.year = "2005"`, `start.month = "01"`,
  `end.year = "2005"`, and `end.month = "03"`.
}
