% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact_single.R
\name{plot_impact_single}
\alias{plot_impact_single}
\title{Plot exposure impact from a single unit}
\usage{
plot_impact_single(
  data.linked,
  data.units,
  link.to = "zips",
  zcta.dataset = NULL,
  counties. = NULL,
  map.month,
  map.unitID,
  plot.name = NULL,
  metric = "N",
  legend.lims = NULL,
  legend.name = NULL,
  graph.dir = NULL,
  zoom = TRUE,
  ...
)
}
\arguments{
\item{data.linked}{data for plotting as output from disperseR::link_all_units()}

\item{data.units}{information on unit locations as output from disperseR::units()}

\item{link.to}{spatial scale for plotting. One of 'zips', 'counties', or 'grids' that should match original input to disperseR::link_all_units()}

\item{zcta.dataset}{ZIP code/ZCTA spatial dataset as imported by
disperseR::get_data() or a subset. Required only if link.to = 'zips'.}

\item{counties.}{County sf object used when link.to = 'counties'.}

\item{map.month}{specific month to map in form YYYYMM if time.agg = 'month'. Can be created using disperseR::get_yearmon()}

\item{map.unitID}{specific unit to map as string. Must match 'ID' column in data.units.}

\item{plot.name}{plot title string. Defaults to NULL, or a descriptive title combining the metric, map.month, and map.unitID}

\item{metric}{plotting metric in `data.linked` input. Defaults to 'hyads', the default output of disperseR::calculate_exposure().}

\item{legend.lims}{legend limits string. Defaults to NULL, which creates a scale from 0 to the 95th percentile of the 'metric'. Should be in the form c(lower,upper).}

\item{legend.name}{legend name string. Defaults to NULL, which creates a legend title of 'Aggregate HyADS Exposure'.}

\item{graph.dir}{location to save output.}

\item{zoom}{Logical. If TRUE, zoom to the data extent; if FALSE, show the
full US extent.}

\item{...}{modeling parameters passed to ggplot2::theme()}
}
\value{
Creates ggplot object of spatial HyADS impacts
}
\description{
`plot_impact_single()` takes output from `link_all_units()` and creates a spatial plot
}
