% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_monthly_links_subfun.R
\name{read_countylinks_subfun}
\alias{read_countylinks_subfun}
\title{Read linked county data from an .fst file}
\usage{
read_countylinks_subfun(i, files)
}
\arguments{
\item{i}{Index into `files`.}

\item{files}{Character vector of .fst paths.}
}
\value{
A data.table of county links for the selected file.
}
\description{
Read linked county data from an .fst file
}
\keyword{internal}
