% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion_subfun.R
\name{read_listing_file}
\alias{read_listing_file}
\title{Read a HYSPLIT listing file}
\usage{
read_listing_file(file_path)
}
\arguments{
\item{file_path}{Path to the `listing` file.}
}
\value{
Character vector of listing lines.
}
\description{
Read a HYSPLIT listing file
}
\keyword{internal}
