% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_monthly_links_subfun.R
\name{read_ziplinks_subfun}
\alias{read_ziplinks_subfun}
\title{Read linked ZIP data from an .fst file}
\usage{
read_ziplinks_subfun(i, files)
}
\arguments{
\item{i}{Index into `files`.}

\item{files}{Character vector of .fst paths.}
}
\value{
A data.table of ZIP links for the selected file.
}
\description{
Read linked ZIP data from an .fst file
}
\keyword{internal}
