% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spatial_area}
\alias{spatial_area}
\title{Calculate area of spatial features}
\usage{
spatial_area(x)
}
\arguments{
\item{x}{sf object or terra SpatVector with polygon geometry}
}
\value{
Numeric vector of areas
}
\description{
Modern replacement for rgeos::gArea.
}
\keyword{internal}
