% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spatial_centroid}
\alias{spatial_centroid}
\title{Get centroids of spatial features}
\usage{
spatial_centroid(x)
}
\arguments{
\item{x}{sf object or terra SpatVector}
}
\value{
Centroids as sf or SpatVector
}
\description{
Modern replacement for rgeos::gCentroid.
}
\keyword{internal}
