% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spatial_union}
\alias{spatial_union}
\title{Union spatial objects}
\usage{
spatial_union(x)
}
\arguments{
\item{x}{sf object or terra SpatVector}
}
\value{
Unified spatial object
}
\description{
Modern replacement for rgeos::gUnion.
}
\keyword{internal}
