% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_all_units_subfun.R
\name{trim_pbl}
\alias{trim_pbl}
\title{Trim parcels below planetary boundary layer}
\usage{
trim_pbl(Min, rasterin)
}
\arguments{
\item{Min}{Input data.table with lon, lat, height, Pdate columns}

\item{rasterin}{PBL height SpatRaster}
}
\value{
Trimmed data.table
}
\description{
Trim parcels below planetary boundary layer
}
\keyword{internal}
