% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{units_monthly}
\alias{units_monthly}
\title{Monthly power plant unit emissions}
\format{
data.table with unit identifiers, year, month, and emissions columns.
}
\usage{
units_monthly
}
\description{
Monthly emissions data by unit, used for exposure weighting.
}
\keyword{datasets}
