% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk data for ZIP and census geographies}
\format{
data.table with ZIP identifiers and associated geography fields.
}
\usage{
crosswalk
}
\description{
Crosswalk table used to link ZIP codes to counties and other
census geographies.
}
\keyword{datasets}
