% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disperseR-package.R
\docType{package}
\name{disperseR-package}
\alias{disperseR-package}
\alias{disperseR}
\title{disperseR: Run HYSPLIT Many Times in Parallel}
\description{
disperseR runs HYSPLIT dispersion models and calculates the HyADS
exposure metric, aggregating results to ZIP code level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/josbyb2020/disperseR-test2026}
  \item Report bugs at \url{https://github.com/josbyb2020/disperseR-test2026/issues}
}

}
\author{
\strong{Maintainer}: Lucas Henneman \email{lhenneman@gmail.com}

Authors:
\itemize{
  \item Christine Choirat
  \item Maja Garbulinska
}

}
