% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_all_units_subfun.R
\name{disperser_link_zips}
\alias{disperser_link_zips}
\title{Link dispersion to ZIP codes}
\usage{
disperser_link_zips(
  month_YYYYMM = NULL,
  start.date = NULL,
  end.date = NULL,
  unit,
  duration.run.hours = 240,
  pbl.height = NULL,
  crosswalk. = NULL,
  res.link. = 12000,
  overwrite = FALSE,
  pbl. = TRUE,
  return.linked.data. = TRUE
)
}
\arguments{
\item{month_YYYYMM}{Month in YYYYMM format}

\item{start.date}{Start date}

\item{end.date}{End date}

\item{unit}{Unit data.table}

\item{duration.run.hours}{Duration in hours}

\item{pbl.height}{PBL height raster}

\item{crosswalk.}{Crosswalk data.table}

\item{res.link.}{Grid resolution}

\item{overwrite}{Overwrite existing files}

\item{pbl.}{Apply PBL normalization}

\item{return.linked.data.}{Return linked data}
}
\value{
data.table with ZIP code links
}
\description{
Link dispersion to ZIP codes
}
