% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{dispersion_read}
\alias{dispersion_read}
\title{Read dispersion CSV outputs from a run directory}
\usage{
dispersion_read(archive_folder)
}
\arguments{
\item{archive_folder}{Directory containing `GIS_part_###_ps.csv` files.}
}
\value{
A data.frame of particle positions with an `hour` column.
}
\description{
Read dispersion CSV outputs from a run directory
}
