% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{get_crs}
\alias{get_crs}
\title{Get CRS from spatial object}
\usage{
get_crs(x)
}
\arguments{
\item{x}{sf object or terra Spat* object}
}
\value{
CRS as character string
}
\description{
Modern replacement for sp::proj4string.
}
