% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{get_crs_definitions}
\alias{get_crs_definitions}
\title{Common CRS definitions}
\usage{
get_crs_definitions()
}
\value{
Named list of CRS strings
}
\description{
Standard CRS definitions used in disperseR.
}
