% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion_subfun.R
\name{hysplit_config_init}
\alias{hysplit_config_init}
\title{Write default HYSPLIT configuration files}
\usage{
hysplit_config_init(dir)
}
\arguments{
\item{dir}{Directory to write `SETUP.CFG` and `ASCDATA.CFG`.}
}
\value{
NULL. Files are written to disk.
}
\description{
Write default HYSPLIT configuration files
}
