% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PP.units.monthly1995_2017}
\alias{PP.units.monthly1995_2017}
\title{Monthly unit emissions (1995-2017)}
\format{
data.table with unit identifiers, year, month, and emissions fields.
}
\usage{
PP.units.monthly1995_2017
}
\description{
Historical monthly emissions data for power plant units.
}
\keyword{datasets}
