% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_monthly_links_subfun.R
\name{read_gridlinks_subfun}
\alias{read_gridlinks_subfun}
\title{Read linked grid data from an .fst file}
\usage{
read_gridlinks_subfun(i, files)
}
\arguments{
\item{i}{Index into `files`.}

\item{files}{Character vector of .fst paths.}
}
\value{
A data.table of grid links for the selected file.
}
\description{
Read linked grid data from an .fst file
}
