% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{read_shapefile}
\alias{read_shapefile}
\title{Read shapefile}
\usage{
read_shapefile(path, ...)
}
\arguments{
\item{path}{Path to shapefile (.shp)}

\item{...}{Additional arguments passed to st_read}
}
\value{
sf object
}
\description{
Modern replacement for rgdal::readOGR and raster::shapefile.
}
