% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spatial_intersect}
\alias{spatial_intersect}
\title{Intersect spatial objects}
\usage{
spatial_intersect(x, y)
}
\arguments{
\item{x}{First sf object or terra SpatVector}

\item{y}{Second sf object or terra SpatVector}
}
\value{
Intersection of x and y
}
\description{
Modern replacement for rgeos::gIntersection.
}
