% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{spatial_join}
\alias{spatial_join}
\title{Spatial join}
\usage{
spatial_join(x, y, join = sf::st_intersects, ...)
}
\arguments{
\item{x}{sf object}

\item{y}{sf object}

\item{join}{Function defining spatial relationship (default: st_intersects)}

\item{...}{Additional arguments}
}
\value{
sf object with joined attributes
}
\description{
Modern replacement for sp::over.
}
