% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel_subfun.R
\name{subset_nc_date}
\alias{subset_nc_date}
\title{Subset NetCDF raster by date}
\usage{
subset_nc_date(hpbl_file = NULL, hpbl_brick = NULL, varname = NULL, vardate)
}
\arguments{
\item{hpbl_file}{Path to NetCDF file (optional)}

\item{hpbl_brick}{SpatRaster object (optional)}

\item{varname}{Variable name in NetCDF}

\item{vardate}{Date to extract}
}
\value{
Single-layer SpatRaster
}
\description{
Extracts a single layer from a multi-layer PBL raster 
based on the specified date.
}
