% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_all_units_subfun.R
\name{trim_zero}
\alias{trim_zero}
\title{Trim parcels with height = 0}
\usage{
trim_zero(Min)
}
\arguments{
\item{Min}{Input data.table}
}
\value{
Trimmed data.table
}
\description{
Trim parcels with height = 0
}
