% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodecoordinate}
\alias{zipcodecoordinate}
\title{ZIP code coordinate reference}
\format{
data.table with ZIP and coordinate columns.
}
\usage{
zipcodecoordinate
}
\description{
Coordinates for ZIP code centroids used in mapping.
}
\keyword{datasets}
